# docs/sa/API_Specification.yaml
openapi: 3.0.3
info:
  title: ServiceDesk CRM API
  description: |
    REST API для системы автоматизации сервисной компании по обслуживанию банкоматов и платёжного оборудования.

    ## Аутентификация
    Все защищённые эндпоинты требуют заголовок `Authorization: Bearer <token>`.
    Токен получается через `POST /auth/login`. Срок действия — 480 минут (8 часов).

    ## Формат ошибок
    ```json
    { "error": "ERROR_CODE", "message": "Описание ошибки", "details": {} }
    ```

    ## Пагинация
    Поддерживается параметрами `?page=1&size=20`.
    Ответ всегда включает `total`, `page`, `size`, `pages`.
  version: "2.0.0"
  contact:
    name: Solution Architect
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: http://localhost/api/v1
    description: Docker Compose (via Nginx)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: TICKET_NOT_FOUND
        message:
          type: string
          example: "Заявка #42 не найдена"
        details:
          type: object
      required: [error, message, details]
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
          description: Текущая страница (начиная с 1)
        size:
          type: integer
          description: Размер страницы
        pages:
          type: integer
          description: Общее количество страниц
      required: [total, page, size, pages]

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
    UserRole:
      type: string
      enum: [admin, sales_mgr, svc_mgr, engineer, manager, warehouse, accountant, director]
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        full_name:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
          description: Массив ролей пользователя (union прав)
        phone:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        is_deleted:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, full_name, roles, is_active, is_deleted, created_at, updated_at]
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
          maxLength: 200
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
          default: ["engineer"]
        password:
          type: string
          minLength: 8
        phone:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
      required: [email, full_name, password]
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
          maxLength: 200
        phone:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        new_password:
          type: string
          minLength: 8
          nullable: true

    # ─── Client ───────────────────────────────────────────────────────────────
...
    WorkTemplateResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        equipment_model_id:
          type: integer
        equipment_model_name:
          type: string
        work_items:
          type: array
          items:
            type: object
            properties:
              order:
                type: integer
              description:
                type: string
            required: [order, description]
        parts:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: integer
              part_name:
                type: string
              quantity:
                type: integer
            required: [part_id, part_name, quantity]
        created_at:
          type: string
          format: date-time
      required: [id, name, equipment_model_id, equipment_model_name, work_items, parts, created_at]
    WorkTemplateCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        equipment_model_id:
          type: integer
        work_items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              order:
                type: integer
              description:
                type: string
            required: [order, description]
        parts:
          type: array
          items:
            type: object
            properties:
              part_id:
                type: integer
              quantity:
                type: integer
                minimum: 1
            required: [part_id, quantity]
      required: [name, equipment_model_id, work_items]

    # ─── Spare Part ───────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Токен отсутствует или истёк
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INVALID_TOKEN
            message: "Токен недействителен"
            details: {}
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: "Требуется роль: admin, svc_mgr"
            details: {}
    NotFound:
      description: Запись не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт уникальности
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessError:
      description: Нарушение бизнес-правила
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [Пользователи]
      summary: Список пользователей
      description: Доступно admin, svc_mgr, director.
      parameters:
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/UserRole'
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Пользователи]
      summary: Создать пользователя
      description: Только admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Пользователи]
      summary: Получить пользователя по ID
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: OK
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Пользователи]
      summary: Обновить пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: OK
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Пользователи]
      summary: Удалить пользователя (soft delete)
      description: Только admin. Нельзя удалить себя.
      responses:
        '204':
          description: Удалён
        '400':
          $ref: '#/components/responses/BusinessError'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════
  # CLIENTS
  # ═══════════════════════════════════════════════════════════════
...
  /work-templates:
    get:
      tags: [Шаблоны работ]
      summary: Список шаблонов
      parameters:
        - in: query
          name: equipment_model_id
          schema:
            type: integer
          description: Фильтр по модели оборудования (для выпадающего списка при создании заявки)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkTemplateResponse'

    post:
      tags: [Шаблоны работ]
      summary: Создать шаблон
      description: admin, svc_mgr. work_items — минимум 1 позиция.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkTemplateCreate'
      responses:
        '201':
          description: Шаблон создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkTemplateResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
  /work-templates/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Шаблоны работ]
      summary: Получить шаблон
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkTemplateResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Шаблоны работ]
      summary: Обновить шаблон
      description: admin, svc_mgr.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkTemplateCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkTemplateResponse'

    delete:
      tags: [Шаблоны работ]
      summary: Удалить (soft delete)
      description: admin, svc_mgr.
      responses:
        '204':
          description: Удалён

  # ═══════════════════════════════════════════════════════════════
  # SPARE PARTS
  # ═══════════════════════════════════════════════════════════════
...